requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
google-api-core==2.24.2
google-api-python-client>=2.0.0
google-auth==2.38.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
openai>=1.0.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rich>=13.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions>=4.0.0
uritemplate==4.1.1
urllib3==2.3.0
email_validator>=2.0.0
